backend/pyproject.toml
[project]
name = "collective-science-backend"
version = "0.1.0"
description = "Shared backend scaffolding for the CollectiveScience MVP"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0,<2.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "pydantic>=2.10.0,<3.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
  "ruff>=0.7.0,<1.0.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
